Module Batrpc_client.Client

Basic client that doesn't have a server component

type 'a with_ctx = Batrpc_core.Handler.ctx * 'a

type t

The client connection

type lifecycle_event

val pp_lifecycle_event : Ppx_deriving_runtime.Format.formatter -> lifecycle_event -> Ppx_deriving_runtime.unit

val show_lifecycle_event : lifecycle_event -> Ppx_deriving_runtime.string

type lifecycle_state

val pp_lifecycle_state : Ppx_deriving_runtime.Format.formatter -> lifecycle_state -> Ppx_deriving_runtime.unit

val show_lifecycle_state : lifecycle_state -> Ppx_deriving_runtime.string

val create : ?buf_pool:Batrpc_core.Buf_pool.t -> ?active:Imandrakit_thread.Switch.t -> ?config:Config.t -> ?on_event:lifecycle_event Imandrakit.Observer.t -> ?track_io_events:bool -> encoding:Batrpc_core.Encoding.t -> timer:Imandrakit_thread.Timer.t -> ic:Batrpc_core.Io.In.t -> oc:Batrpc_core.Io.Out.t -> unit -> t

Create a new client.

parameter on_event
observer for events

parameter track_io_events
if true, instrument ic and oc to record all reads and writes into the on_event observer

val close_and_join : t -> unit

close_and_join rpc closes the connection and waits for the background thread to finish

val close_without_joining : t -> unit

close_without_joining rpc closes the connection but immediately returns.

val active : t -> Imandrakit_thread.Switch.t option

val state : t -> State.t

State for the client end of this connection

val on_event : t -> lifecycle_event Imandrakit.Observer.t

val lifecycle_state : t -> lifecycle_state

val on_close : t -> unit Imandrakit_thread.Fut.t

Future to wait for this to close

val wait_block_close : t -> unit

wait_block_close rpc is Fut.wait_block_exn @@ on_close rpc. Be careful with deadlocks.

exception Closed

val call : t -> ?headers:Batrpc_core.Meta.header list -> ?timeout_s:float -> ('req, Batrpc_core.Service.Value_mode.unary, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Client.rpc -> 'req -> 'res Imandrakit_thread.Fut.t

raises Closed
if the connection is closed

val call_client_stream : t -> ?headers:Batrpc_core.Meta.header list -> ?timeout_s:float -> ('req, Batrpc_core.Service.Value_mode.stream, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Client.rpc -> 'req Batrpc_core.Push_stream.t * 'res Imandrakit_thread.Fut.t

val call_server_stream : t -> ?headers:Batrpc_core.Meta.header list -> ?timeout_s:float -> ('req, Batrpc_core.Service.Value_mode.unary, 'item, Batrpc_core.Service.Value_mode.stream) Batrpc_core.Service.Client.rpc -> init:(unit with_ctx -> 'state) -> on_item:('state -> 'item -> unit) -> on_close:('state -> 'res with_ctx) -> 'req -> 'res Imandrakit_thread.Fut.t