Module Batrpc_client.State

type 'a with_ctx = Batrpc_core.Handler.ctx * 'a

type t

Client-side state. This tracks the state of in-flight requests and active streams.

val create : ?middlewares:Middleware.t list -> default_timeout_s:float -> framing_config:Batrpc_core.Framing.config -> encoding:Batrpc_core.Encoding.t -> unit -> t

New client state

val add_middleware : t -> Middleware.t -> unit

val handle_response : t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> unit -> unit

val handle_error : t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> unit -> unit

val handle_stream_item : t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> unit -> unit

val handle_stream_close : t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> unit -> unit

val send_heartbeat : t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> unit -> unit

Send a heartbeat message

val call : t -> timer:Imandrakit_thread.Timer.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> ?headers:Batrpc_core.Meta.header list -> ?timeout_s:float -> ('req, Batrpc_core.Service.Value_mode.unary, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Client.rpc -> 'req -> 'res Imandrakit_thread.Fut.t

Call a regular RPC method.

val call_client_stream : t -> timer:Imandrakit_thread.Timer.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> ?headers:Batrpc_core.Meta.header list -> ?timeout_s:float -> ('req, Batrpc_core.Service.Value_mode.stream, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Client.rpc -> 'req Batrpc_core.Push_stream.t * 'res Imandrakit_thread.Fut.t

Call a RPC method that takes a client stream. This returns a pair stream, res; you can then push items onto the stream using Push_stream.push. When it's done, call Push_stream.close and the future will resolve once the server has processed it.

val call_server_stream : t -> timer:Imandrakit_thread.Timer.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> ?headers:Batrpc_core.Meta.header list -> ?timeout_s:float -> ('req, Batrpc_core.Service.Value_mode.unary, 'item, Batrpc_core.Service.Value_mode.stream) Batrpc_core.Service.Client.rpc -> init:(unit with_ctx -> 'state) -> on_item:('state -> 'item -> unit) -> on_close:('state -> 'res with_ctx) -> 'req -> 'res Imandrakit_thread.Fut.t

Call a RPC method that takes a request and returns a stream from the server.