Module Batrpc_server.For_client

Basic server for a single client, on an existing connection.

type 'a with_ctx = Batrpc_core.Handler.ctx * 'a

type t

val create : ?buf_pool:Batrpc_core.Buf_pool.t -> ?active:Imandrakit_thread.Switch.t -> ?state:State.t -> ?config:Config.t -> encoding:Batrpc_core.Encoding.t -> runner:Batrpc_core.Runner.t -> timer:Imandrakit_thread.Timer.t -> ic:Batrpc_core.Io.In.t -> oc:Batrpc_core.Io.Out.t -> unit -> t

type handler = State.handler

val close : t -> unit

close rpc closes the connection. Thread-safe.

val runner : t -> Moonpool.Runner.t

The background runner for tasks

val active : t -> Imandrakit_thread.Switch.t option

val state : t -> State.t

State for the server end of this connection

val add_service : t -> handler Batrpc_core.Service.Server.t -> unit

Add some server-side service to this connection

val on_close : t -> unit Imandrakit_thread.Fut.t

Future to wait for this to close

val run : t -> unit

run rpc runs the server loops on the current thread, blocking it, until close rpc is called.