Module Batrpc_server.State

Server side for a given connection.

type t

Server side for a given connection to a client

val pp : Stdlib.Format.formatter -> t -> unit

val show : t -> string

type handler

type 'a with_ctx = Batrpc_core.Handler.ctx * 'a

val mk_handler : ('req, Batrpc_core.Service.Value_mode.unary, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Server.rpc -> ('req -> 'res Imandrakit_thread.Fut.t) -> handler

Make a simple request/reply handler

val mk_handler_full : ('req, Batrpc_core.Service.Value_mode.unary, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Server.rpc -> ('req with_ctx -> 'res with_ctx Imandrakit_thread.Fut.t) -> handler

Make a full request/reply handler, with headers

val mk_client_stream_handler : ('req, Batrpc_core.Service.Value_mode.stream, 'res, Batrpc_core.Service.Value_mode.unary) Batrpc_core.Service.Server.rpc -> init:(unit with_ctx -> 'state) -> on_item:('state -> 'req -> unit) -> on_close:('state -> 'res with_ctx) -> unit -> handler

Make a handler for client-stream requests. The handler is a state machine that reacts to new stream elements and returns a response upon stream closure.

val mk_server_stream_handler : ('req, Batrpc_core.Service.Value_mode.unary, 'res, Batrpc_core.Service.Value_mode.stream) Batrpc_core.Service.Server.rpc -> ('req with_ctx -> 'res Batrpc_core.Push_stream.t -> unit) -> handler

Make a handler for server-stream. The handler can push values back to the client. The request ends when the stream is closed by the handler.

val create : ?middlewares:Middleware.t list -> services:handler Batrpc_core.Service.Server.t list -> framing_config:Batrpc_core.Framing.config -> unit -> t

val add_middleware : t -> Middleware.t -> unit

val add_service : t -> handler Batrpc_core.Service.Server.t -> unit

val list_services : t -> handler Batrpc_core.Service.Server.t list

val find_meth : t -> string -> (string * handler) option

val send_heartbeat : t -> encoding:Batrpc_core.Encoding.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> unit -> unit

Send a heartbeat message

val handle_request : t -> encoding:Batrpc_core.Encoding.t -> runner:Batrpc_core.Runner.t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> unit -> unit

We have read, from ic, the metadata for a new message of type Meta.Request. Handle this by running (in the current thread) the corresponding handler and reply to it.

parameter runner
where to spawn the handler for this request?

val handle_stream_item : t -> encoding:Batrpc_core.Encoding.t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> unit -> unit

Handle stream close, in the current thread. We must preserve the order of processing.

val handle_stream_close : t -> encoding:Batrpc_core.Encoding.t -> meta:Batrpc_core.Meta.meta -> ic:Batrpc_core.Io.In.t -> oc:Batrpc_core.Io.Out.t Imandrakit_thread.Lock.t -> unit -> unit

Handle stream close, in the current thread. We must preserve the order of processing.