Module Batrpc_unix.Simple_timer

Basic timer.

Runs in a background thread.

include module type of struct include Batrpc_core.Timer end

type t = Batrpc_core.Timer.t = {

run_after_s : float -> (unit -> unit) -> unit;

run_every_s : float -> (unit -> unit) -> unit;

terminate : unit -> unit;

}

Timer, to run callbacks in the future

exception Stop_timer

val run_after_s : t -> float -> (unit -> unit) -> unit

run_after_s t f waits t seconds and then runs f

val run_every_s : t -> float -> (unit -> unit) -> unit

run_every ~initial t f waits initial seconds and then runs f() every t seconds. f () can raise Stop_timer to stop the loop.

val terminate : t -> unit

val create : unit -> t