Module Batrpc_unix.Tcp_server

TCP server.

The server will spawn a thread per client connection, and use an executor to handle individual requests from each client.

type t

val create : ?server_state:Batrpc_core.Server_state.t -> ?on_new_client:(Batrpc_core.Rpc_conn.t -> Unix.sockaddr -> unit) -> ?config_socket:(Unix.file_descr -> unit) -> ?reuseaddr:bool -> ?middlewares:Batrpc_core.Middleware.Server.t list -> active:Batrpc_core.Switch.t -> runner:Batrpc_unix.Util_.Runner.t -> timer:Batrpc_core.Timer.t -> services: Batrpc_core.Server_state.handler Batrpc_unix.Util_.Service.Server.t list -> Unix.sockaddr -> t Batrpc_core.Error.result

Create and bind the TCP server.

parameter executor
used to handle individual requests.

parameter config_socket
custom function to set options on server socket before it's used, but after basic options and reuseaddr have been set.

val add_middleware : t -> Batrpc_core.Middleware.Server.t -> unit

val active : t -> Batrpc_core.Switch.t

val run : t -> unit

Run the server in current thread. This blocks the current thread for potentially a very long time.

val terminate : t -> unit

Terminate server.